# docs/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "description" : "Household task management API.\n\nConventions callers need:\n- Every response is the envelope `{ success, message, data }`.\n- `user`, `group` and `task` are addressed by opaque `uid`; every other resource by numeric `id`.\n- Timestamps are UTC with no zone suffix — parse as UTC, convert for display.\n- Actors are derived from the token; never send a user or member id to identify the caller.\n- Errors: 400 validation, 401 unauthenticated, 403 forbidden, 404 not found, 409 duplicate or violated business rule, 429 rate limited (see `Retry-After`).\n- Role requirements are stated per operation and are not modelled in this spec.\n",
    "title" : "TasKlean API",
    "version" : "0.0.1-SNAPSHOT"
  },
  "servers" : [ {
    "url" : "/",
    "description" : "Relative to the host serving this API"
  } ],
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
...
  "paths" : {
...
    "/api/devices" : {
      "get" : {
        "description" : "Lists a user's active devices.",
        "operationId" : "getDevicesByUser",
        "parameters" : [ {
          "description" : "the user id",
          "in" : "query",
          "name" : "userId",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseListDeviceResponse"
                }
              }
            },
            "description" : "<code>200 OK</code> with the devices"
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthorized"
          },
          "403" : {
            "$ref" : "#/components/responses/Forbidden"
          },
          "429" : {
            "$ref" : "#/components/responses/TooManyRequests"
          },
          "500" : {
            "$ref" : "#/components/responses/InternalServerError"
          }
        },
        "summary" : "Lists a user's active devices.",
        "tags" : [ "Devices" ]
      },
      "post" : {
        "description" : "Registers a device for push notifications.",
        "operationId" : "registerDevice",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/DeviceRequest"
              }
            }
          },
          "description" : "the device details (token, type, user)",
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseDeviceResponse"
                }
              }
            },
            "description" : "<code>201 Created</code> with the registered device"
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthorized"
          },
          "403" : {
            "$ref" : "#/components/responses/Forbidden"
          },
          "429" : {
            "$ref" : "#/components/responses/TooManyRequests"
          },
          "500" : {
            "$ref" : "#/components/responses/InternalServerError"
          }
        },
        "summary" : "Registers a device for push notifications.",
        "tags" : [ "Devices" ]
      }
    },
    "/api/devices/{id}" : {
      "delete" : {
        "description" : "Soft-deletes (deactivates) a device.",
        "operationId" : "deactivateDevice",
        "parameters" : [ {
          "description" : "the device id",
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description" : "<code>200 OK</code> with a confirmation message"
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthorized"
          },
          "403" : {
            "$ref" : "#/components/responses/Forbidden"
          },
          "429" : {
            "$ref" : "#/components/responses/TooManyRequests"
          },
          "500" : {
            "$ref" : "#/components/responses/InternalServerError"
          }
        },
        "summary" : "Soft-deletes (deactivates) a device.",
        "tags" : [ "Devices" ]
      },
      "get" : {
        "description" : "Fetches a device by id.",
        "operationId" : "getDevice",
        "parameters" : [ {
          "description" : "the device id",
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseDeviceResponse"
                }
              }
            },
            "description" : "<code>200 OK</code> with the device"
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthorized"
          },
          "403" : {
            "$ref" : "#/components/responses/Forbidden"
          },
          "429" : {
            "$ref" : "#/components/responses/TooManyRequests"
          },
          "500" : {
            "$ref" : "#/components/responses/InternalServerError"
          }
        },
        "summary" : "Fetches a device by id.",
        "tags" : [ "Devices" ]
      },
      "put" : {
        "description" : "Updates a device's descriptive fields.",
        "operationId" : "updateDevice",
        "parameters" : [ {
          "description" : "the device id",
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/DeviceRequest"
              }
            }
          },
          "description" : "the new device details",
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseDeviceResponse"
                }
              }
            },
            "description" : "<code>200 OK</code> with the updated device"
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthorized"
          },
          "403" : {
            "$ref" : "#/components/responses/Forbidden"
          },
          "429" : {
            "$ref" : "#/components/responses/TooManyRequests"
          },
          "500" : {
            "$ref" : "#/components/responses/InternalServerError"
          }
        },
        "summary" : "Updates a device's descriptive fields.",
        "tags" : [ "Devices" ]
      }
    },
...
  },
  "components" : {
    "responses" : {
      "BadRequest" : {
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description" : "Validation failure, malformed body, or a bad path/query parameter"
      },
      "Forbidden" : {
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description" : "Authenticated, but lacking the platform or group role this operation requires"
      },
      "InternalServerError" : {
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description" : "Unexpected server error"
      },
      "TooManyRequests" : {
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description" : "Rate limit exceeded — see the Retry-After header"
      },
      "Unauthorized" : {
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description" : "No valid access token — or, on the public auth endpoints, rejected credentials"
      }
    },
    "schemas" : {
...
      "ApiResponseDeviceResponse" : {
        "type" : "object",
        "description" : "Standard response envelope wrapping every API result with a <code>success</code> flag,\n an optional <code>message</code>, and the payload <code>data</code>. Construct via the static\n factory methods rather than the builder directly.",
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/DeviceResponse"
          },
          "message" : {
            "type" : "string"
          },
          "success" : {
            "type" : "boolean"
          }
        }
      },
...
      "ApiResponseListDeviceResponse" : {
        "type" : "object",
        "description" : "Standard response envelope wrapping every API result with a <code>success</code> flag,\n an optional <code>message</code>, and the payload <code>data</code>. Construct via the static\n factory methods rather than the builder directly.",
        "properties" : {
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/DeviceResponse"
            }
          },
          "message" : {
            "type" : "string"
          },
          "success" : {
            "type" : "boolean"
          }
        }
      },
...
      "ApiResponseVoid" : {
        "type" : "object",
        "description" : "Standard response envelope wrapping every API result with a <code>success</code> flag,\n an optional <code>message</code>, and the payload <code>data</code>. Construct via the static\n factory methods rather than the builder directly.",
        "properties" : {
          "data" : { },
          "message" : {
            "type" : "string"
          },
          "success" : {
            "type" : "boolean"
          }
        }
      },
...
      "DeviceRequest" : {
        "type" : "object",
        "properties" : {
          "appVersion" : {
            "type" : "string"
          },
          "browserInfo" : {
            "type" : "string"
          },
          "deviceName" : {
            "type" : "string"
          },
          "deviceToken" : {
            "type" : "string",
            "minLength" : 1
          },
          "deviceType" : {
            "type" : "string",
            "minLength" : 1
          },
          "osVersion" : {
            "type" : "string"
          },
          "userId" : {
            "type" : "integer",
            "format" : "int64"
          }
        },
        "required" : [ "deviceToken", "deviceType", "userId" ]
      },
      "DeviceResponse" : {
        "type" : "object",
        "properties" : {
          "appVersion" : {
            "type" : "string"
          },
          "browserInfo" : {
            "type" : "string"
          },
          "dateCreated" : {
            "type" : "string",
            "format" : "date-time"
          },
          "dateUpdated" : {
            "type" : "string",
            "format" : "date-time"
          },
          "deviceName" : {
            "type" : "string"
          },
          "deviceToken" : {
            "type" : "string"
          },
          "deviceType" : {
            "type" : "string"
          },
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "isActive" : {
            "type" : "boolean"
          },
          "lastUsed" : {
            "type" : "string",
            "format" : "date-time"
          },
          "osVersion" : {
            "type" : "string"
          },
          "userId" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
      "ErrorResponse" : {
        "type" : "object",
        "description" : "The standard envelope, as returned for a failed request.",
        "properties" : {
          "data" : {
            "type" : "object",
            "description" : "Always null on an error"
          },
          "message" : {
            "type" : "string",
            "description" : "Human-readable error message"
          },
          "success" : {
            "type" : "boolean",
            "default" : false
          }
        }
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "description" : "Access token from /api/auth/login, /api/auth/verify-email, /api/auth/google or /api/auth/refresh.",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  }
}